Added tunnels
I added MapFactory, an implementation of the FactoryMethod pattern, which is responsible for creating maps. I also added the class TunnelMap, which extends the class GlobeMap because it differs only in that it stores tunnels. I placed the tunnels in a regular list in a class attribute, not in MapField, because I assumed that MapField is responsible for actions at a position, and not for movement, so a tunnel there doesn't fit. Animals are moved by the Map using the move method and when they enter a tunnel, they disappear from the map and reappear on the other side after exiting the tunnel.