Added visual representation for MapElements
Given that different MapElements have their own representation on the map, which will expand to include images, texts, backgrounds, I have placed all this in the VisualRepresentation class. Additionally, I added the Builder pattern to this class since each class can have its own implementation of appearance, which will vary in the number of arguments. Furthermore, I added the drawing of the tunnel on the map, represented as the border of a square. I didn’t use StrokeWidth because it breaks the appearance of other squares on the map, instead, I used two nested squares.