Added counting descendants in O(1) time
When we click on an animal, function collects count of all descendants and marks all descendants as descendant of an alfaAnimal.
Then each time the descendant reproduces we increase counter in alfaAnimal object.
First step - counting all descendants can be complex so we run a new thread not to freeze gui.